Recreated TriangleShape, edited Polygon classes.
The TriangleShape class was edited to make it a subclass of
PolygonShape. PolygonShape's Polygon field was edited to be protected
visibility, so the TriangleShape class can make use of that field.
Furthermore, when a Polygon only has three vertices, the
convexDecomposition class will now only return a list with the current
Polygon instance as its only element.